Move test to ext_test.go
